reinstall nodes when deleting policy
Razor 0.15.0 adds functionality to delete policies. Previously
ASM simply disabled them and reinstalled the nodes associated
with those policies. It appeared that simply deleting the
policies would take the place of those steps but that is not the
case. When you delete a policy but do not reinstall attached
nodes the node is left in a bad state. When the node reboots
razor thinks it has an O/S already installed on it but does not
know what it is. It ends up booting the "noop" task which means
that it just boots into the hard disk as opposed to booting the
microkernel like it should.

This commit calls reinstall-node on all nodes attached to a
policy after deleting the policy to fix this issue.